Version2/app/requirements.txt
ultralytics
numpy
opencv-python-headless
pandas
Pillow
pyYAML
requests
tqdm
streamlit
streamlit-chat
bytetracker
streamlit_extras